Harden schema integrity: active-enrollment guard, Decimal marks, join-table tenancy
Three findings from a fresh full-schema re-audit, one theme: make the
database enforce what the product promises.

- One ACTIVE enrollment per student (migration f4b6d8a0c2e4).
  uq_enrollment_student_year only guards within a year; nothing stopped
  a student holding ACTIVE rows in two different years, and the
  lifecycle service resolves the current enrollment with
  scalar_one_or_none() — a duplicate would surface as a 500 on the
  student page and a blocked promotion. A partial unique index
  (student_id WHERE status = 'ACTIVE') closes it, the same pattern as
  uq_one_active_year_per_school. Live data pre-checked: zero
  duplicates. The new test proves both directions — a second ACTIVE
  row across years is rejected while the closed row promotion writes
  is allowed.

- Marks annotations stop lying (no migration; columns were already
  Numeric). Ten Mapped[float] annotations on Numeric columns across
  examination, mastery, question_bank, question_paper and report_card
  become Mapped[Decimal] — the same annotation-vs-runtime mismatch
  that produced the historical fee bug, now eliminated for academic
  marks, which are as authoritative as money.

- Defense-in-depth tenancy on join tables (migration a6c8e0b2d4f6).
  student_parent_map, notice_read_receipts and student_transport were
  scoped only through JOINs to tenant-owned parents — the same gap the
  notifications table had (war-room f31). Each now carries its own
  non-null, indexed school_id FK, backfilled from the unambiguous
  parent inside one transactional migration. Every writer sets it:
  notice mark-read, transport assignment, link_parent, the test
  fixture and six seed scripts.

Also fixes the pre-existing import-sort error in question_bank.py,
since the Decimal import touches that block.

Verified: both migrations upgrade/downgrade/upgrade cleanly; 121 tests
pass across thirteen suites (enrollments, lifecycle, promotion,
academic, portal, notifications, notices, exam questions, answer-sheet
eval, mastery API and compute, report card, tenant isolation); ruff
clean on all touched files; model imports resolve.

+++ apps/api/app/db/models/examination.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import enum
 import uuid
 from datetime import date
+from decimal import Decimal
 
 from sqlalchemy import (
     Boolean,
@@ -50,7 +51,7 @@ class Exam(BaseModel):
     )
     exam_type: Mapped[ExamType] = mapped_column(Enum(ExamType), nullable=False)
     title: Mapped[str] = mapped_column(String(200), nullable=False)
-    total_marks: Mapped[float] = mapped_column(Numeric(6, 2), nullable=False)
+    total_marks: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
     date: Mapped[date | None] = mapped_column(Date)
     # Whole-exam chapter/topic tag — slip/unit tests map 1:1 to a chapter.
     topic: Mapped[str | None] = mapped_column(String(120))
@@ -79,7 +80,7 @@ class ExamMark(BaseModel):
     student_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("students.id"), nullable=False
     )
-    marks_obtained: Mapped[float] = mapped_column(Numeric(6, 2), nullable=False)
+    marks_obtained: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
     # {"1": 3.5, "2": 4} question-no -> marks; absent key = unattempted (internal
     # choice). When present, marks_obtained is server-derived as the sum.
     question_marks: Mapped[dict | None] = mapped_column(JSONB)
